chore: rename directory from room-list to rooms
