+++ src/pages/comparison/RoomComparison.tsx
