+++ src/pages/comparison/styled.ts
