+++ src/pages/index.tsx
@@ -1,16 +1,16 @@
 import React from 'react';
 import { Provider } from 'react-redux';
 import { BrowserRouter, Route, Switch } from 'react-router-dom';
+import { createStore } from 'store';
 
-import { createStore } from '../store';
 import Login from './auth/Login';
 import ChecklistPage from './checklist/ChecklistPage';
-import RoomList from './list/RoomList';
+import RoomComparisonPage from './comparison/RoomComparison';
 import MainPage from './main/MainPage';
 import PersonaAnalysisPage from './persona/PersonaAnalysis';
 import PersonaAnalysisResultPage from './persona/PersonaAnalysisResult';
 import PersonaQuestionPage from './question/PersonaQuestion';
-import RoomComparisonPage from './room/RoomComparison';;
+import RoomList from './rooms/RoomList';
 
 export default function EntryRoute() {
   const store = createStore();
@@ -23,9 +23,9 @@ export default function EntryRoute() {
           <Route path='/persona/question' component={PersonaQuestionPage} />
           <Route path='/persona' component={PersonaAnalysisPage} />
           <Route path='/checklist' component={ChecklistPage} />
-          <Route path='/room' component={RoomComparisonPage} />
+          <Route path='/comparison' component={RoomComparisonPage} />
           <Route path='/auth' component={Login} />
-          <Route path='/room-list' component={RoomList} />
+          <Route path='/rooms' component={RoomList} />
           <Route path='/' component={MainPage} />
         </Switch>
       </BrowserRouter>
