+++ src/pages/rooms/styled.ts
